requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-decouple
PyMySQL
django-cors-headers
redis
rest-framework-simplejwt
PyJWT
psycopg2
django-redis
channels
channels-redis